+++ CHANGELOG.md
@@ -111,7 +111,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 - Deprecated `DataModule` properties: `train_transforms`, `val_transforms`, `test_transforms`, `size`, `dims` ([#8851](https://github.com/PyTorchLightning/pytorch-lightning/pull/8851))
 
 
--
+- Deprecated `add_to_queue`, `get_from_queue` from Lightning Module([9118](https://github.com/PyTorchLightning/pytorch-lightning/pull/9118))
 
 
 -